refactor(api): moved actions in ES to a dedicated file
